# onebrain-web/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "BrandThemeOut": {
        "properties": {
          "accent_color": {
            "title": "Accent Color",
            "type": "string"
          },
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "app_id": {
            "default": "",
            "title": "App Id",
            "type": "string"
          },
          "background_color": {
            "title": "Background Color",
            "type": "string"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "danger_color": {
            "title": "Danger Color",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "logo_url": {
            "default": "",
            "title": "Logo Url",
            "type": "string"
          },
          "muted_color": {
            "title": "Muted Color",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "secondary_color": {
            "title": "Secondary Color",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "success_color": {
            "title": "Success Color",
            "type": "string"
          },
          "surface_color": {
            "title": "Surface Color",
            "type": "string"
          },
          "text_color": {
            "title": "Text Color",
            "type": "string"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          },
          "warning_color": {
            "title": "Warning Color",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "primary_color",
          "secondary_color",
          "accent_color",
          "background_color",
          "surface_color",
          "text_color",
          "muted_color",
          "success_color",
          "warning_color",
          "danger_color"
        ],
        "title": "BrandThemeOut",
        "type": "object"
      },
      "BundleOut": {
        "properties": {
          "apps": {
            "items": {
              "type": "string"
            },
            "title": "Apps",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "spaces": {
            "items": {
              "type": "string"
            },
            "title": "Spaces",
            "type": "array"
          }
        },
        "required": [
          "id",
          "label",
          "description",
          "spaces",
          "apps",
          "modules"
        ],
        "title": "BundleOut",
        "type": "object"
      },
...
      "CustomerProvisionCreate": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "account_kind": {
            "default": "organization",
            "pattern": "^(person|organization|family|project)$",
            "title": "Account Kind",
            "type": "string"
          },
          "app_brand_themes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/app__routers__provisioning__BrandThemeInput"
            },
            "title": "App Brand Themes",
            "type": "object"
          },
          "brand_theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__routers__provisioning__BrandThemeInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "bundle_id": {
            "default": "full_stack",
            "maxLength": 80,
            "title": "Bundle Id",
            "type": "string"
          },
          "current_migration": {
            "default": "",
            "maxLength": 80,
            "title": "Current Migration",
            "type": "string"
          },
          "customer_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Customer Name",
            "type": "string"
          },
          "deployment_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Id"
          },
          "deployment_type": {
            "default": "dedicated_railway",
            "maxLength": 80,
            "title": "Deployment Type",
            "type": "string"
          },
          "initial_version": {
            "default": "0.1.0",
            "maxLength": 80,
            "minLength": 1,
            "title": "Initial Version",
            "type": "string"
          },
          "mint_integration_keys": {
            "default": true,
            "title": "Mint Integration Keys",
            "type": "boolean"
          },
          "module_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Module Versions",
            "type": "object"
          },
          "region": {
            "default": "",
            "maxLength": 80,
            "title": "Region",
            "type": "string"
          },
          "release_ring": {
            "default": "manual",
            "maxLength": 80,
            "title": "Release Ring",
            "type": "string"
          }
        },
        "required": [
          "customer_name"
        ],
        "title": "CustomerProvisionCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvisionedAccountOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_id": {
            "default": "",
            "title": "Owner User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "title": "ProvisionedAccountOut",
        "type": "object"
      },
      "ProvisionedAppOut": {
        "properties": {
          "allowed_purposes": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Purposes",
            "type": "array"
          },
          "app_id": {
            "title": "App Id",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "enabled_space_ids": {
            "items": {
              "type": "string"
            },
            "title": "Enabled Space Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "app_id",
          "enabled_space_ids",
          "allowed_purposes"
        ],
        "title": "ProvisionedAppOut",
        "type": "object"
      },
      "ProvisionedCredentialOut": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "app_id": {
            "title": "App Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "purposes": {
            "items": {
              "type": "string"
            },
            "title": "Purposes",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "space_ids": {
            "items": {
              "type": "string"
            },
            "title": "Space Ids",
            "type": "array"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "tenant_id",
          "account_id",
          "app_id",
          "label",
          "scopes",
          "space_ids",
          "purposes"
        ],
        "title": "ProvisionedCredentialOut",
        "type": "object"
      },
      "ProvisionedDeploymentOut": {
        "properties": {
          "current_migration": {
            "default": "",
            "title": "Current Migration",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "string"
          },
          "customer_name": {
            "title": "Customer Name",
            "type": "string"
          },
          "deployment_type": {
            "title": "Deployment Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "region": {
            "default": "",
            "title": "Region",
            "type": "string"
          },
          "release_ring": {
            "title": "Release Ring",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customer_name",
          "deployment_type",
          "release_ring",
          "current_version"
        ],
        "title": "ProvisionedDeploymentOut",
        "type": "object"
      },
      "ProvisionedModuleOut": {
        "properties": {
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "module_id",
          "version",
          "status"
        ],
        "title": "ProvisionedModuleOut",
        "type": "object"
      },
      "ProvisionedSpaceOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "title": "ProvisionedSpaceOut",
        "type": "object"
      },
      "ProvisioningResultOut": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/ProvisionedAccountOut"
          },
          "app_brand_themes": {
            "items": {
              "$ref": "#/components/schemas/BrandThemeOut"
            },
            "title": "App Brand Themes",
            "type": "array"
          },
          "apps": {
            "items": {
              "$ref": "#/components/schemas/ProvisionedAppOut"
            },
            "title": "Apps",
            "type": "array"
          },
          "brand_theme": {
            "$ref": "#/components/schemas/BrandThemeOut"
          },
          "bundle_id": {
            "title": "Bundle Id",
            "type": "string"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/ProvisionedCredentialOut"
            },
            "title": "Credentials",
            "type": "array"
          },
          "deployment": {
            "$ref": "#/components/schemas/ProvisionedDeploymentOut"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ProvisionedModuleOut"
            },
            "title": "Modules",
            "type": "array"
          },
          "spaces": {
            "items": {
              "$ref": "#/components/schemas/ProvisionedSpaceOut"
            },
            "title": "Spaces",
            "type": "array"
          }
        },
        "required": [
          "bundle_id",
          "account",
          "spaces",
          "apps",
          "deployment",
          "modules",
          "brand_theme"
        ],
        "title": "ProvisioningResultOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__routers__provisioning__BrandThemeInput": {
        "properties": {
          "accent_color": {
            "default": "#a66e2f",
            "maxLength": 7,
            "title": "Accent Color",
            "type": "string"
          },
          "background_color": {
            "default": "#f4f2ee",
            "maxLength": 7,
            "title": "Background Color",
            "type": "string"
          },
          "danger_color": {
            "default": "#b4453e",
            "maxLength": 7,
            "title": "Danger Color",
            "type": "string"
          },
          "logo_url": {
            "default": "",
            "maxLength": 500,
            "title": "Logo Url",
            "type": "string"
          },
          "muted_color": {
            "default": "#5f6671",
            "maxLength": 7,
            "title": "Muted Color",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "default": "#16191e",
            "maxLength": 7,
            "title": "Primary Color",
            "type": "string"
          },
          "secondary_color": {
            "default": "#3e5573",
            "maxLength": 7,
            "title": "Secondary Color",
            "type": "string"
          },
          "success_color": {
            "default": "#1f7a4d",
            "maxLength": 7,
            "title": "Success Color",
            "type": "string"
          },
          "surface_color": {
            "default": "#ffffff",
            "maxLength": 7,
            "title": "Surface Color",
            "type": "string"
          },
          "text_color": {
            "default": "#101828",
            "maxLength": 7,
            "title": "Text Color",
            "type": "string"
          },
          "warning_color": {
            "default": "#b98a4e",
            "maxLength": 7,
            "title": "Warning Color",
            "type": "string"
          }
        },
        "title": "BrandThemeInput",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "onebrain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/provisioning/bundles": {
      "get": {
        "operationId": "list_bundles_api_provisioning_bundles_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundleOut"
                  },
                  "title": "Response List Bundles Api Provisioning Bundles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bundles",
        "tags": [
          "provisioning"
        ]
      }
    },
    "/api/provisioning/customers": {
      "post": {
        "operationId": "provision_customer_api_provisioning_customers_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerProvisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provision Customer",
        "tags": [
          "provisioning"
        ]
      }
    },
...
  }
}